rag_service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
openai>=1.3.0
numpy>=1.24.0
pymysql>=1.1.0
cryptography>=41.0.0
chromadb>=0.4.15

